Rename route files
* Renamed the route files, much simpler, no need to name them based on middleware
