+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The complete changelog for the Costs to Expect REST API, follows the format defined at https://keepachangelog.com/en/1.0.0/
 
+## [v1.12.0] - 2019-04-xx
+### Changed
+- Renamed the route files, filenames relate to access, not the middleware that runs.
+
 ## [v1.11.0] - 2019-04-02
 ### Added 
 - All messages and strings (route descriptions, fields in OPTIONS requests, error messages) localised.
