+++ routes/api.php
@@ -1,9 +1,6 @@
 <?php
 
-require_once ('api/auth.php');
-require_once ('api/auth-api_convert-route-params.php');
-require_once ('api/convert-route-params_convert-get-params_log-requests.php');
-require_once ('api/convert-route-params_log-requests.php');
-require_once ('api/convert-get-params_log-requests.php');
-require_once ('api/log-requests.php');
-require_once ('api/no-middleware.php');
+require_once('api/auth.php');
+require_once('api/private-routes.php');
+require_once('api/public-routes.php');
+require_once('api/public-summary-routes.php');
